config files refactored
